+++ R/fractional.computations.R
@@ -1851,8 +1851,11 @@ rSPDE.construct.matern.loglike <- function(object, Y, A,
                                  return_negative_likelihood = TRUE,
                                  pivot = TRUE){
         param_vector <- likelihood_process_inputs(user_kappa, user_sigma, user_nu, sigma.e)
-
-        loglik <- function(theta){
+        
+        loglik <- function(theta, obj = NULL){
+          if(!is.null(obj)){
+            object <- obj
+          }
           if(is.null(user_kappa)){
           kappa <- likelihood_process_parameters(theta = theta, 
                   param_vector = param_vector,